support for storing prefixes
We employ a special bit mask value that conceptually indicates a
bit coming before all real bits in each byte: any comparison with
it will direct the search to the right child.

This mainly changes the insertion code, which needs to detect and
insert prefix nodes, and avoid assuming a null terminator stored
after the keys.

The search code also needs a subtle change. Originally, all
comparisons would add "virtual null bytes" after the key to
direct short keys to the left child:

    c = 0;
    if (q->byte < ulen) {
        c = ubytes[q->byte];
    }
    direction = (1 + (q->otherbits | c)) >> 8;

But this would give a wrong result in a prefix node, since its
mask would direct even zero bytes to the right child. To avoid
checking for prefix nodes all the time, we simply change it to
use a default direction instead:

    direction = 0;
    if (q->byte < ulen) {
        c = ubytes[q->byte];
        direction = (1 + (q->otherbits | c)) >> 8;
    }